Refactor QLDB logic into similar style - New cleaned up ledger removes need for backwards-compatibility
