+++ package.json
@@ -7,7 +7,7 @@
       "email": "nicholas.para@dw.com"
     }
   ],
-  "version": "2.0.0",
+  "version": "2.0.1",
   "private": true,
   "dependencies": {
     "@types/formidable": "^1.2.2",
