--- src/qldb-Constants.ts
@@ -1,3 +1,0 @@
-export const LEDGER_NAME = 'deptoolkit';
-export const DOC_TABLE_NAME = 'Document';
-export const DOC_INDEX_KEY = 'sku';
